Move stipend calculation to clarify_and_take_passed_resources
