+++ basic_bootloader/src/bootloader/supported_ees.rs
@@ -1,8 +1,8 @@
 use crate::bootloader::EVM_EE_BYTE;
 use errors::{EESubsystemError, InterfaceError};
 use zk_ee::{
-    execution_environment_type::ExecutionEnvironmentType, interface_error,
-    memory::slice_vec::SliceVec, system::*, wrap_error,
+    common_structs::CalleeAccountProperties, execution_environment_type::ExecutionEnvironmentType,
+    interface_error, memory::slice_vec::SliceVec, system::*, wrap_error,
 };
 
 #[allow(type_alias_bounds)]
@@ -36,12 +36,16 @@ impl<'ee, S: EthereumLikeTypes> SupportedEEVMState<'ee, S> {
         ee_version: ExecutionEnvironmentType,
         resources_available_in_caller_frame: &mut S::Resources,
         desired_ergs_to_pass: Ergs,
+        call_request: &ExternalCallRequest<S>,
+        callee_parameters: &CalleeAccountProperties,
     ) -> Result<S::Resources, EESubsystemError> {
         match ee_version {
             ExecutionEnvironmentType::EVM => {
                 SystemBoundEVMInterpreter::<S>::clarify_and_take_passed_resources(
                     resources_available_in_caller_frame,
                     desired_ergs_to_pass,
+                    call_request,
+                    callee_parameters,
                 )
                 .map_err(wrap_error!())
             }
